Add Update & Delete Action
On the category-list page when the user clicks on edit, the user is redirected to update category page with the user id passed as params. The Update Category component then uses the id passed to first fetch the category that needs to be updated then populate it in the input section and then the user can update or delete the title according to their needs.