Add Comments To Posts
Created and added the  'AddComment' component to the post details page. The post id was passed as a prop in the component. The post Id along with the description entered by the user was passed to the createPostAction slice. Using axios post action the comment was stored in the database.