CommentList & DeleteCommentAction
Added another component to the PostDetails Page i.e. CommentList to display the comments of each post.
Normally we create a fetch Action for displaying something but this time changes were made in the backend folder.  Comments were virtually added to the Post Model and then populated into the allPosts & postDetails controller. The controllers allowed comments to be fetched from the database and displayed along with other post details.
The comments for each post were passed as prop to the CommentList component and then using map the array of comments was rendered on the PostDetails Page.
The Delete functionality was also enabled by creating  delete action and dispatching it with each comment. The comment Id was passed along to the action to delete the comment using axios.